feat: implement multi-homeroom teacher support and add automated attendance reporting and monitoring commands

+++ database/migrations/2026_07_20_000000_add_wali_kelas_2_id_to_kelas_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('kelas', function (Blueprint $table) {
-            $table->unsignedBigInteger('wali_kelas_2_id')->nullable()->after('wali_kelas_id');
+            $table->unsignedInteger('wali_kelas_2_id')->nullable()->after('wali_kelas_id');
             $table->foreign('wali_kelas_2_id')->references('id')->on('guru')->onDelete('set null');
         });
     }
